web3 dropdown -> web3 buttons
