+++ client/styles/components/login_web3_buttons.scss
@@ -1,0 +1,14 @@
+@import 'client/styles/shared';
+
+.LoginWeb3Buttons {
+    .login-with-web3 {
+        .cui-icon {
+            position: relative;
+            top: -2px;
+            margin-left: 4px;
+            svg polyline {
+                stroke-width: 2.25px;
+            }
+        }
+    }
+}
