--- client/styles/components/login_with_wallet_dropdown.scss
@@ -1,25 +1,0 @@
-@import 'client/styles/shared';
-
-.LoginWithWalletDropdown {
-}
-
-.login-with-web3 {
-    .cui-icon {
-        position: relative;
-        top: -2px;
-        margin-left: 4px;
-        svg polyline {
-            stroke-width: 2.25px;
-        }
-    }
-}
-
-.LoginWithWalletModal {
-    .chain-login-label {
-        display: flex;
-        .chain-login-label-name {
-            margin: 3px 0;
-            margin-left: 10px;
-        }
-    }
-}
